Splice billing table store
