+++ src/stores/Tables/Billing/Hydrator.tsx
@@ -2,6 +2,7 @@ import { PostgrestFilterBuilder } from '@supabase/postgrest-js';
 import useBillingHistory from 'components/tables/Billing/useBillingHistory';
 import { useZustandStore } from 'context/Zustand/provider';
 import { useEffect } from 'react';
+import { useUnmount } from 'react-use';
 import { SelectTableStoreNames } from 'stores/names';
 import { useBilling_hydrateContinuously } from 'stores/Tables/Billing/hooks';
 import {
@@ -24,6 +25,11 @@ export const BillingHistoryTableHydrator = ({
         SelectableTableStore['setQuery']
     >(SelectTableStoreNames.BILLING, selectableTableStoreSelectors.query.set);
 
+    const resetState = useZustandStore<
+        SelectableTableStore,
+        SelectableTableStore['resetState']
+    >(SelectTableStoreNames.BILLING, selectableTableStoreSelectors.state.reset);
+
     const hydrateState = useBilling_hydrateContinuously();
 
     const { billingHistory, error, isValidating } = useBillingHistory({
@@ -39,6 +45,11 @@ export const BillingHistoryTableHydrator = ({
             hydrateState(billingHistory, error);
         }
     }, [hydrateState, billingHistory, error, isValidating]);
+
+    // Reset state when leaving until we work out how we want to cache table stuff
+    useUnmount(() => {
+        resetState();
+    });
 
     // eslint-disable-next-line react/jsx-no-useless-fragment
     return <>{children}</>;
